new graph methods
testing two new graphing methods and cleaned up files
