+++ README.md
@@ -17,7 +17,7 @@ Approach (1):
 (The larger the difference between radius of the circle on the flat plane and the radius x, the further below or above the plane is relative to the object.)
 ___________________________________________________________________________________________________________________________________________________________________________________
 
-Approach (2):
+Approach (2) - active:
 - Approximate pair of spheres with smallest overlap by determining pair of spheres with smallest overlap along vector from one center to the other
 - Calculate center of overlap along vector from one center to the other center
 - Calculate radius of circle where spheres intersect plane normal to vector
@@ -41,3 +41,20 @@ display?
     - scatter plot of points contained in spheres (on request)
 
 ___________________________________________________________________________________________________________________________________________________________________________________
+
+Issues / to change
+
+Sometimes fails to find points within intersection area when area becomes small.
+considering changing approach instead of increasing the number of samples.
+ - does this only occur when the radius of instersection and overlap distance are very different?
+    - use radius and overlap distance to create cube bounding point generation instead of generating a sphere encompassing the larger of the two
+        - could greatly cut down on the total area the points have to cover and therefore increase the likelihood of generated points being within intersection
+
+ - generate a lower ammount of points and retry with larger ammounts if too few points are found?
+    - still pretty brute force, should be a better solution
+
+Clean up terminal view or add GUI (I dont like making gui's)
+
+Actually go through and comment the code
+
+Further adjust plot and maybe find a way to open it in a dedicated window
