+++ archive/sphereIntersectTest.py
