+++ main.py
@@ -4,240 +4,267 @@ import pyperclip
 from global_hotkeys import *
 import os
 import matplotlib.pyplot as plt
+import plotly.graph_objects as go
 
 def_rad = 200
-plotResolution = .2 #points per meters squared
-plotExtension = 5 #factor to increase radius when generating plot points
+samples = 500000
+maxPlotPoints = 20000
+#factor for adjusting random distribution in sphere scattering
+#k>3 will favour outer edge of sphere
+k = 3
 
-EPS = 1e-9
+#def plot_shader(plotPoints):
 
-
-def _violation_and_argmax(x, spheres):
-    """Return (max_violation, idx) for point x with respect to spheres.
-    violation for sphere i is d(x, c_i) - r_i. Feasible if max_violation <= 0.
-    """
-    x = np.asarray(x, dtype=float)
-    centers = np.array([s[:3] for s in spheres], dtype=float)
-    radii = np.array([s[3] for s in spheres], dtype=float)
-    dists = np.linalg.norm(centers - x, axis=1)
-    viol = dists - radii
-    k = int(np.argmax(viol))
-    return float(viol[k]), k
-
-
-def _aabb_intersection(spheres):
-    """Axis-aligned bounding box of the intersection of all sphere AABBs.
-    Returns (lo, hi) where lo/hi are 3-vectors. If empty, returns (None, None).
-    """
-    centers = np.array([s[:3] for s in spheres], dtype=float)
-    radii = np.array([s[3] for s in spheres], dtype=float)
-    lo = np.max(centers - radii[:, None], axis=0)
-    hi = np.min(centers + radii[:, None], axis=0)
-    if np.any(lo > hi):
-        return None, None
-    return lo, hi
-
-
-def _random_points_in_box(lo, hi, n):
-    return lo + (hi - lo) * np.random.rand(n, 3)
-
-
-def _chebyshev_center(spheres, x0=None, iters=500, step0=0.25):
-    """Approximate Chebyshev center of the intersection (center of the largest
-    inscribed ball). Uses subgradient ascent on g(x) = min_i (r_i - ||x-c_i||).
-    Returns (x, inner_radius_estimate, max_violation).
-    """
-    centers = np.array([s[:3] for s in spheres], dtype=float)
-    if x0 is None:
-        x = centers.mean(axis=0)
-    else:
-        x = np.array(x0, dtype=float)
-
-    best_x = x.copy()
-    best_val = -np.inf
-
-    for t in range(1, iters + 1):
-        # Identify the most "active" (tightest) sphere
-        dists = np.linalg.norm(centers - x, axis=1)
-        margins = np.array([s[3] for s in spheres], dtype=float) - dists
-        k = int(np.argmin(margins))  # tightest constraint
-        val = margins[k]
-        if val > best_val:
-            best_val = val
-            best_x = x.copy()
-
-        # Subgradient step to increase the tightest margin
-        if dists[k] < EPS:
-            # If exactly at a center, nudge randomly
-            direction = np.random.randn(3)
-            direction /= (np.linalg.norm(direction) + EPS)
-        else:
-            direction = (x - centers[k]) / dists[k]
-
-        step = step0 / np.sqrt(t)
-        x = x + step * direction
-
-    # Final stats at best_x
-    viol, _ = _violation_and_argmax(best_x, spheres)
-    inner_r = -viol  # if positive, it's the radius of largest contained ball
-    return best_x, inner_r, viol
-
-def common_sphere_intersection(
-    spheres,
-    samples=50000,
-    refine_iters=600,
-    tolerance=1e-6,
-    return_points=False,
-    random_seed=None,
-):
-    """
-    Approximate the common intersection region of multiple spheres.
-
-    Parameters
-    ----------
-    spheres : list[(x,y,z,r)]
-        Sphere centers and radii.
-    samples : int
-        Number of random samples for a quick feasible point search.
-    refine_iters : int
-        Iterations for Chebyshev-center refinement.
-    tolerance : float
-        Numerical tolerance on feasibility (<= 0 means strictly feasible).
-    return_points : bool
-        If True, also return a down-sampled set of points lying in the
-        intersection (useful for visualization).
-    random_seed : int or None
-        RNG seed for reproducibility.
-
-    Returns
-    -------
-    dict with keys:
-        'feasible' : bool
-        'center' : (x,y,z) of approximate Chebyshev center (if feasible)
-        'inner_radius' : radius of the largest ball guaranteed inside the
-            intersection (>= 0 when feasible; 0 when tangency)
-        'max_violation' : positive value means infeasible by that margin
-        'aabb' : (lo, hi) bounding box used for sampling (if any)
-        'points' : (N,3) array of feasible sample points (if return_points)
-    """
-    rng = np.random.default_rng(random_seed)
-    spheres = [(float(x), float(y), float(z), float(r)) for (x, y, z, r) in spheres]
-
-    lo, hi = _aabb_intersection(spheres)
-    result = {
-        'feasible': False,
-        'center': None,
-        'inner_radius': 0.0,
-        'max_violation': None,
-        'aabb': (lo, hi),
-    }
-
-    if lo is None:
-        result['max_violation'] = np.inf
-        return result
-
-    # Quick sampling to find a feasible point (if one exists)
-    P = _random_points_in_box(lo, hi, samples) if samples > 0 else np.empty((0, 3))
-    if samples > 0:
-        centers = np.array([s[:3] for s in spheres], dtype=float)
-        radii = np.array([s[3] for s in spheres], dtype=float)
-        diffs = P[:, None, :] - centers[None, :, :]
-        dists = np.linalg.norm(diffs, axis=2)
-        inside = (dists <= radii[None, :] + tolerance).all(axis=1)
-        feasible_points = P[inside]
-    else:
-        feasible_points = np.empty((0, 3))
-
-    x0 = feasible_points.mean(axis=0) if feasible_points.size else np.mean([s[:3] for s in spheres], axis=0)
-
-    # Refine Chebyshev center (works whether or not we found a feasible point)
-    x_star, inner_r, viol = _chebyshev_center(spheres, x0=x0, iters=refine_iters)
-
-    result.update({
-        'center': tuple(map(float, x_star)),
-        'inner_radius': float(max(0.0, inner_r)),
-        'max_violation': float(max(0.0, viol)),
-        'feasible': bool(viol <= tolerance),
-    })
-
-    if return_points:
-        # Down-sample feasible points for plotting if too many
-        if len(feasible_points) > 2000:
-            idx = rng.choice(len(feasible_points), size=2000, replace=False)
-            feasible_points = feasible_points[idx]
-            result['points'] = feasible_points
-
-    return result
-
-def plotPoints(plotPoints, center):
+def plot_points(plotPoints):
     fig = plt.figure()
     ax = fig.add_subplot(projection='3d')
-    ax.scatter(xs = plotPoints['x'], ys = plotPoints['z'], zs = plotPoints['y'], c='r', marker='o')
+    for point in plotPoints[:maxPoints]:
+        ax.scatter(xs = point[0], ys = point[1], zs = point[2], c='r', marker='o')
     ax.set_xlabel('X')
     ax.set_ylabel('Z')
     ax.set_zlabel('Y')
     ax.set_autoscale_on(True)
     plt.show()
 
-def scatterPoints(spheres, center, radius):
-    spheres = [(float(x), float(y), float(z), float(r)) for (x, y, z, r) in spheres]
+def plot_scatter3d(points, max_points=200_000):
+    points = np.array(points)
+    n = len(points)
     
-    generationRange = radius*plotExtension
-    samples = (2*int(generationRange*plotResolution))**3
-    testPoints = np.array([center[0] - generationRange + 2*generationRange*np.random.rand(samples), 
-                  center[1] - generationRange + 2*generationRange*np.random.rand(samples), 
-                  center[2] - generationRange + 2*generationRange*np.random.rand(samples)]).T
-    validxVals = []
-    validyVals = []
-    validzVals = []
-    for point in testPoints:
+    if n > max_points:
+        idx = np.random.choice(n, size=max_points, replace=False)
+        points = points[idx]
+    
+    fig = go.Figure(data=[go.Scatter3d(
+        x=points[:, 0],
+        y=points[:, 1],
+        z=points[:, 2],
+        mode='markers',
+        marker=dict(size=2, opacity=0.5)
+    )])
+    
+    fig.update_layout(
+        scene=dict(
+            xaxis_title='X',
+            yaxis_title='Y',
+            zaxis_title='Z'
+        ),
+        title=f"3D Scatter ({len(points):,} points shown)"
+    )
+    fig.show()
+
+def plot_volume(points, bins=50):
+    points = np.array(points)
+    H, edges = np.histogramdd(points, bins=bins)
+    
+    # Build voxel grid
+    x_centers = (edges[0][:-1] + edges[0][1:]) / 2
+    y_centers = (edges[1][:-1] + edges[1][1:]) / 2
+    z_centers = (edges[2][:-1] + edges[2][1:]) / 2
+    
+    X, Y, Z = np.meshgrid(x_centers, y_centers, z_centers, indexing="ij")
+    
+    fig = go.Figure(data=go.Volume(
+        x=X.flatten(),
+        y=Y.flatten(),
+        z=Z.flatten(),
+        value=H.flatten(),
+        isomin=1,          # min cutoff (ignore empty voxels)
+        isomax=H.max(),    # max density
+        opacity=0.1,       # transparency
+        surface_count=15   # number of isosurfaces
+    ))
+    
+    fig.update_layout(
+        scene=dict(
+            xaxis_title='X',
+            yaxis_title='Y',
+            zaxis_title='Z'
+        ),
+        title=f"3D Volume Rendering (bins={bins})"
+    )
+    fig.show()
+
+def sphere_intersection(spheres):
+    scatterSamples = None
+    #calculate vector between center of spheres
+    #consider reworking this to only check new pairs against old best when sphere added
+    bestOverlap = []
+    spherePairs = [(s1, s2) for s1 in spheres for s2 in spheres if s1 != s2]
+    for spherePair in spherePairs:
+        overlap = spherePair[0][3] + spherePair[1][3] - vector_magnitude(vector(spherePair[0], spherePair[1]))
+        try:
+            if bestOverlap[1] > overlap:
+                bestOverlap = [spherePair, overlap]
+        except:
+            bestOverlap = [spherePair, overlap]
+    
+    v_u = vector(bestOverlap[0][0], bestOverlap[0][1])
+    distance = vector_magnitude(v_u)
+    v_u = [x / distance for x in v_u]
+    distCenter = (distance**2 + bestOverlap[0][0][3]**2 - (bestOverlap[0][1][3]**2)) / (2*distance)
+    vCenter = [x * distCenter for x in v_u]
+    centerPoint = [bestOverlap[0][0][0] + vCenter[0], bestOverlap[0][0][1] + vCenter[0], bestOverlap[0][0][2] + vCenter[0]]
+
+    intersectionRadius = math.sqrt(bestOverlap[0][0][3]**2 - distCenter**2)
+
+    if len(spheres) > 2:
+        if bestOverlap[1] <= intersectionRadius:
+            scatterSamples = sphere_scatter(centerPoint, intersectionRadius, samples)
+        else:
+            print("primary overlap distance greater than intersection radius")
+            scatterSamples = sphere_scatter(centerPoint, bestOverlap[1], samples)
+        scatterSamples = valid_points(spheres, scatterSamples)
+        if len(scatterSamples) > 0:
+            scatterCenter = scatter_center(scatterSamples)
+            scatterBounding = find_scatter_bounds(scatterSamples, scatterCenter)
+            print("Center of scatter: ", scatterCenter)
+            print(scatterBounding)
+        else:
+            print("center of intersection: ", [round(i) for i in centerPoint])
+            print("intersection radius: ", round(intersectionRadius))
+            print("intersection overlap: ", round(bestOverlap[1]))
+            roundedVector = [f"{num:.2f}" for num in v_u]
+            print("direction vector: ", roundedVector)
+    else:
+        print("center of intersection: ", [round(i) for i in centerPoint])
+        print("intersection radius: ", round(intersectionRadius))
+        print("intersection overlap: ", round(bestOverlap[1]))
+        roundedVector = [f"{num:.2f}" for num in v_u]
+        print("direction vector: ", roundedVector)
+    return(scatterSamples)
+    
+def valid_points(spheres, scatterSamples):
+    validPoints = []
+    for point in scatterSamples:
         valid = True
         for sphere in spheres:
-            dist = math.sqrt(((point[0]-sphere[0])**2) + ((point[1]-sphere[1])**2) + ((point[2]-sphere[2])**2))
-            if dist > sphere[3]:
+            if vector_magnitude(vector(sphere, point)) > sphere[3]:
                 valid = False
-                break
         if valid:
-            validxVals.append(point[0])
-            validyVals.append(point[1])
-            validzVals.append(point[2])
-    return {'x': validxVals, 'y': validyVals, 'z': validzVals}
+            validPoints.append(point)
+    return validPoints
+
+def update_list(spheres, newSphere, removedSpheres):
+    updatedList = []
+    addNewSphere = True
+    for sphere in spheres:
+        if vector_magnitude(vector(sphere, newSphere)) + newSphere[3] <= sphere[3]:
+            if sphere[3] <= newSphere[3]:
+                updatedList.append(sphere)
+                addNewSphere = False
+        else:
+            updatedList.append(sphere)
+    if addNewSphere == True:
+        updatedList.append(newSphere)
+    else:
+        removedSpheres.append(newSphere)
+
+    return [updatedList, removedSpheres]
+
+def scatter_center(scatterSamples):
+    xavg = 0
+    yavg = 0
+    zavg = 0
+
+    for point in scatterSamples:
+        xavg += point[0]
+        yavg += point[1]
+        zavg += point[2]
+    xavg = xavg/len(scatterSamples)
+    yavg = yavg/len(scatterSamples)
+    zavg = zavg/len(scatterSamples)
+    return [int(xavg), int(yavg), int(zavg)]
+
+def vector_magnitude(v):
+    mag = math.sqrt(v[0]**2 + v[1]**2 + v[2]**2)
+    return mag
+
+def vector(sphereOne, sphereTwo):
+    v = [sphereTwo[0] - sphereOne[0], sphereTwo[1] - sphereOne[1], sphereTwo[2] - sphereOne[2]]
+    return v
+
+def sphere_scatter(center, radius, samples):
+    v = np.random.normal(0, 1, (samples, 3))
+    v /= np.linalg.norm(v, axis = 1)[:, np.newaxis]
+    rad = np.random.rand(samples) ** (1/k) * radius
+    points = v * rad[:, np.newaxis]
+    points = points + center
+    return points
+
+def find_scatter_bounds(scatterPoints, center):
+    scatterPoints = (np.array(scatterPoints) - np.array(center))
+    xbounds = np.array([0, 0])
+    ybounds = np.array([0, 0])
+    zbounds = np.array([0, 0])
+    for point in scatterPoints:
+        if point[0] > xbounds[1]:
+            xbounds[1] = point[0]
+        if point[0] < xbounds[0]:
+            xbounds[0] = point[0]
+
+        if point[1] > ybounds[1]:
+            ybounds[1] = point[1]
+        if point[1] < ybounds[0]:
+            ybounds[0] = point[1]
+
+        if point[2] > zbounds[1]:
+            zbounds[1] = point[2]
+        if point[2] < zbounds[0]:
+            zbounds[0] = point[2]
+    #return {"x":xbounds.astype(int).tolist(), "y": ybounds.astype(int).tolist(), "z": zbounds.astype(int).tolist()}
+    return {"x": (xbounds + center[0]).astype(int).tolist(), "y": (ybounds + center[1]).astype(int).tolist(), "z": (zbounds + center[2]).astype(int).tolist()}
+
+
 
 while True:
     os.system('cls')
-    points = []
-    result = None
+    sphereList = []
+    removedSpheres = []
+    scatterSamples = None
+
     while True:
         print("enter a point from clip board or enter 'reset' to reset")
         userInput = input()
-        if userInput == "reset":
+        if userInput == "r":
             break
-        elif userInput == "plot":
-            if result is not None:
-                plotPoints(scatterPoints(points, center = result['center'], radius = result['inner_radius']), result['center'])
+        elif userInput == "p":
+            if scatterSamples is not None:
+                plot_scatter3d(scatterSamples, )
             else:
-                print("Please add more points")
+                print("no points exist for plotting")
+        elif userInput == "u":
+            try:
+                removedSpheres.append(sphereList.pop())
+            except:
+                print("no sphere to remove")
         else:
             try:
                 rad = int(userInput)
             except:
                 rad = def_rad
+
+            ############
                 
             os.system('cls')
             inputBuff = pyperclip.paste().split()[6:9:1]
-            inputBuffFloat = [float(item) for item in inputBuff]
-            inputBuffFloat.append(rad)
-            points.append(inputBuffFloat)
+            newSphere = [float(item) for item in inputBuff]
+            newSphere.append(rad)
 
-            print("entered points:")
-            for item in points:
-                print(item)
+            [sphereList, removedSpheres] = update_list(sphereList, newSphere, removedSpheres)
+            
+            if len(removedSpheres) >= 1:
+                print("inactive spheres")
+                for sphere in removedSpheres:
+                    print(sphere)
+                print("\n")
 
-            if len(points) > 1:
-                result = common_sphere_intersection(
-                    points, samples = 20000, refine_iters = 1200, return_points = False)
-                print("possible locations:")
-                print("center: ", result["center"], "radius: ", result['inner_radius'])
+            print("active spheres:")
+            for sphere in sphereList:
+                print(sphere)
+            print("\n")
+            if len(sphereList) >=2:
+                scatterSamples = sphere_intersection(sphereList)
+
